# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "SignuSever API",
    "description": "A simple API for SignuApp",
    "contact": {
      "email": "sobrenombre@gmail.com"
    }
  },
  "host": "localhost:3000",
  "schemes": [
    "http",
    "https"
  ],
  "basePath": "/api/",
  "paths": {
...
    "/pdfs/{pdf_id}": {
      "get": {
        "tags": [
          "pdf"
        ],
        "summary": "Download a PDF by ID",
        "description": "Download a PDF if user is logged and is owner or signer",
        "parameters": [
          {
            "name": "pdf_id",
            "in": "path",
            "description": "ID of PDF to download",
            "required": true,
            "type": "string",
            "format": "ObjectId"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "pdf"
        ],
        "summary": "Delete a PDF by ID",
        "description": "Delete a PDF by ID if I am logged and I am the owner",
        "parameters": [
          {
            "name": "pdf_id",
            "in": "path",
            "description": "ID of PDF to delete",
            "required": true,
            "type": "string",
            "format": "ObjectId"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "pdf"
        ],
        "summary": "Sign a PDF by ID",
        "description": "Sign a PDF by ID if I am logged and I am a signer",
        "parameters": [
          {
            "name": "pdf_id",
            "in": "path",
            "description": "ID of PDF to delete",
            "required": true,
            "type": "string",
            "format": "ObjectId"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/pdfs/status/{pdf_id}": {
      "get": {
        "tags": [
          "pdf"
        ],
        "summary": "Get info of a PDF",
        "description": "Get info of a PDF",
        "parameters": [
          {
            "name": "pdf_id",
            "in": "path",
            "description": "ID of PDF to download",
            "required": true,
            "type": "string",
            "format": "ObjectId"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "pdf": {
                      "$ref": "#/definitions/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/pdfs/unlock/{pdf_id}": {
      "put": {
        "tags": [
          "pdf"
        ],
        "summary": "Unlock a PDF",
        "description": "Unlock a PDF allowing to sign it during 60 sec",
        "parameters": [
          {
            "name": "pdf_id",
            "in": "path",
            "description": "ID of PDF to unlock",
            "required": true,
            "type": "string",
            "format": "ObjectId"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "pdf": {
                      "$ref": "#/definitions/Pdf"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/pdfs/addsigner/{pdf_id}": {
      "put": {
        "tags": [
          "pdf"
        ],
        "summary": "Add a new signer to PDF",
        "description": "Add a new signer to PDF if this has no stamp",
        "parameters": [
          {
            "name": "pdf_id",
            "in": "path",
            "description": "ID of PDF to edit",
            "required": true,
            "type": "string",
            "format": "ObjectId"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "pdf": {
                      "$ref": "#/definitions/Pdf"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "user_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://localhost:3000/api/users/login",
      "flow": "implicit"
    }
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "format": "ObjectId"
        },
        "name": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "description": "an existing email"
        },
        "last_edition_date": {
          "type": "string",
          "format": "Date"
        },
        "creation_date": {
          "type": "string",
          "format": "Date"
        },
        "pdfs_owned": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "pdfs_to_sign": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "pdfs_signed": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "activation": {
          "type": "object",
          "properties": {
            "is_activated": {
              "type": "boolean"
            },
            "when": {
              "type": "string"
            }
          }
        },
        "next_email": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string",
              "description": "an existing email"
            },
            "when": {
              "type": "string",
              "format": "Date"
            }
          }
        }
      },
      "xml": {
        "name": "User"
      }
    },
...
  }
}